fix(bigint): retarget int4/order assumptions for bigint surrogate IDs (TRA-720)
Surfaced by Task 20 full-suite verification on the new bigint stack:

* storage/reports.go (6×) and storage/locations.go (1×): change explicit
  `::int[]` casts on id-column ANY() predicates to `::bigint[]`. Without
  the change, pgx tries to encode a `[]int` containing Feistel-permuted
  bigint ids into _int4 (OID 1007) and the row request 500s with
  "is greater than maximum value for int4".

* handlers/{assets,locations}/patch_round_trip_integration_test.go: the
  TagsSetEqualityEcho tests SELECTed persisted tags `ORDER BY id` and
  asserted `seeded[i] == persisted[i]`. Insertion order no longer
  matches id order under Feistel permutation; rewrite both assertions
  to look up persisted rows by id via a map.

* contract-tests/explicit_error_cases.py: the deterministic too_large
  probe used `parent_id=9999999999` to trip the int32 cap. With the cap
  gone (TRA-720), the surviving numeric-max path is `?limit=>1000`;
  retarget the probe at GET /assets?limit=99999 so the
  FieldErrorCode coverage gate still observes too_large.

Also picks up `gofmt` whitespace touch-ups to apispec/postprocess*.go
that `just backend lint` applied during this run.

After these fixes:
  - go test ./... PASS
  - just backend test-integration ./... PASS (was 4 failures)
  - just test-contract PASS (was failing on FieldErrorCode coverage)
  - Feistel parity test PASS
  - MQTT cascade smoke: 1 org → 1 location, 1 asset, 1 asset_scan

+++ backend/internal/storage/reports.go
@@ -117,7 +117,7 @@ func (s *Storage) CountCurrentLocations(ctx context.Context, orgID int, filter r
 		FROM latest_scans ls
 		JOIN trakrf.assets    a ON a.id = ls.asset_id AND a.org_id = $1 AND ` + temporallyEffective("a") + `
 		LEFT JOIN trakrf.locations l ON l.id = ls.location_id AND l.org_id = $1 AND l.deleted_at IS NULL AND ` + temporallyEffective("l") + `
-		WHERE ($2::int[]  IS NULL OR l.id           = ANY($2::int[]))
+		WHERE ($2::bigint[]  IS NULL OR l.id           = ANY($2::bigint[]))
 		  AND ($3::text[] IS NULL OR l.external_key = ANY($3::text[]))
 		  AND ($4::text IS NULL OR a.name ILIKE $4 OR a.external_key ILIKE $4
 			   OR EXISTS (
@@ -125,7 +125,7 @@ func (s *Storage) CountCurrentLocations(ctx context.Context, orgID int, filter r
 				   WHERE ai.asset_id = a.id AND ai.is_active = true AND ai.deleted_at IS NULL AND ` + temporallyEffective("ai") + ` AND ai.value ILIKE $4
 			   ))
 		  AND (a.deleted_at IS NULL OR $5::bool)
-		  AND ($6::int[]  IS NULL OR a.id           = ANY($6::int[]))
+		  AND ($6::bigint[]  IS NULL OR a.id           = ANY($6::bigint[]))
 		  AND ($7::text[] IS NULL OR a.external_key = ANY($7::text[]))
 	`
 
@@ -204,7 +204,7 @@ func buildCurrentLocationsQueryDistinctOn(orderBy string) string {
 		FROM latest_scans ls
 		JOIN trakrf.assets a ON a.id = ls.asset_id AND a.org_id = $1 AND ` + temporallyEffective("a") + `
 		LEFT JOIN trakrf.locations l ON l.id = ls.location_id AND l.org_id = $1 AND l.deleted_at IS NULL AND ` + temporallyEffective("l") + `
-		WHERE ($2::int[]  IS NULL OR l.id           = ANY($2::int[]))
+		WHERE ($2::bigint[]  IS NULL OR l.id           = ANY($2::bigint[]))
 		  AND ($3::text[] IS NULL OR l.external_key = ANY($3::text[]))
 		  AND ($4::text IS NULL OR a.name ILIKE $4 OR a.external_key ILIKE $4
 			   OR EXISTS (
@@ -212,7 +212,7 @@ func buildCurrentLocationsQueryDistinctOn(orderBy string) string {
 				   WHERE ai.asset_id = a.id AND ai.is_active = true AND ai.deleted_at IS NULL AND ` + temporallyEffective("ai") + ` AND ai.value ILIKE $4
 			   ))
 		  AND (a.deleted_at IS NULL OR $7::bool)
-		  AND ($8::int[]  IS NULL OR a.id           = ANY($8::int[]))
+		  AND ($8::bigint[]  IS NULL OR a.id           = ANY($8::bigint[]))
 		  AND ($9::text[] IS NULL OR a.external_key = ANY($9::text[]))
 		ORDER BY ` + orderBy + `
 		LIMIT $5 OFFSET $6
@@ -242,7 +242,7 @@ func buildCurrentLocationsQueryTimescale(orderBy string) string {
 		FROM latest_scans ls
 		JOIN trakrf.assets a ON a.id = ls.asset_id AND a.org_id = $1 AND ` + temporallyEffective("a") + `
 		LEFT JOIN trakrf.locations l ON l.id = ls.location_id AND l.org_id = $1 AND l.deleted_at IS NULL AND ` + temporallyEffective("l") + `
-		WHERE ($2::int[]  IS NULL OR l.id           = ANY($2::int[]))
+		WHERE ($2::bigint[]  IS NULL OR l.id           = ANY($2::bigint[]))
 		  AND ($3::text[] IS NULL OR l.external_key = ANY($3::text[]))
 		  AND ($4::text IS NULL OR a.name ILIKE $4 OR a.external_key ILIKE $4
 			   OR EXISTS (
@@ -250,7 +250,7 @@ func buildCurrentLocationsQueryTimescale(orderBy string) string {
 				   WHERE ai.asset_id = a.id AND ai.is_active = true AND ai.deleted_at IS NULL AND ` + temporallyEffective("ai") + ` AND ai.value ILIKE $4
 			   ))
 		  AND (a.deleted_at IS NULL OR $7::bool)
-		  AND ($8::int[]  IS NULL OR a.id           = ANY($8::int[]))
+		  AND ($8::bigint[]  IS NULL OR a.id           = ANY($8::bigint[]))
 		  AND ($9::text[] IS NULL OR a.external_key = ANY($9::text[]))
 		ORDER BY ` + orderBy + `
 		LIMIT $5 OFFSET $6
